consumer/requirements.txt
paho-mqtt
datadog
ddtrace
